Use source.include to pass source files
In order to prevent issues with the length of command-line arguments for
large projects, this change uses the source.include key in the config
file in order to pass the input files to JSDoc.

This change also fixes an issue where every test shared the same mutable
instance of the config object.

Fixes #25